+++ .gitignore
@@ -1,0 +1,29 @@
+# Node modules
+node_modules/
+
+# Build output
+/dist
+
+# Vite build cache
+.vite/
+
+# Logs
+npm-debug.log*
+yarn-debug.log*
+
+# Environment files
+.env
+.env.local
+.env.*.local
+
+# Editor directories
+.vscode/
+.idea/
+
+# OS generated files
+.DS_Store
+Thumbs.db
+
+# Coverage directory
+coverage/
+
