+++ src/App.jsx
@@ -1,59 +1,6 @@
-import React, { useEffect } from 'react'
-import { Routes, Route, useNavigate } from 'react-router-dom'
-import { getMessaging, onMessage } from 'firebase/messaging'
-import PrivateRoute from './components/PrivateRoute'
-import Home from './pages/Home'
-import Login from './pages/Login'
-import Consultorio from './pages/Consultorio' 
-import { useNotification } from './services/NotificationContext'
-import './app.css'; 
+import React from 'react';
+import AppRoutes from './routes/AppRoutes.jsx';
 
-function App() {
-  const navigate = useNavigate();
-  const { notificacao, setNotificacao } = useNotification()
-
-  useEffect(() => {
-    const messaging = getMessaging();
-    const unsubscribe = onMessage(messaging, (payload) => {
-      console.log('Mensagem recebida no foreground:', payload);
-      const data = payload.data || {};
-      const usuarioId = localStorage.getItem('usuarioId');
-      
-      if (data.tipo === 'consulta' && data.meetingId) {
-        setNotificacao({
-          title: payload.notification?.title || 'Nova reunião',
-          meetingId: data.meetingId,
-          userId: usuarioId,
-        })
-      } else {
-        alert(`Nova notificação: ${payload.notification?.title || 'Sem título'}`);
-      }
-    });
-
-    return () => unsubscribe();
-  }, [navigate]);
-
-  return (
-    <Routes>
-      <Route path="/" element={<Login />} />
-      <Route
-        path="/consulta/:meetingId/:userId"
-        element={
-          <PrivateRoute>
-            <Consultorio />
-          </PrivateRoute>
-        }
-      />
-      <Route
-        path="/home"
-        element={
-          <PrivateRoute>
-            <Home />
-          </PrivateRoute>
-        }
-      />
-    </Routes>
-  )
+export default function App() {
+  return <AppRoutes />;
 }
-
-export default App
