+++ src/components/PrivateRoute.jsx
@@ -3,5 +3,5 @@ import { Navigate } from 'react-router-dom'
 
 export default function PrivateRoute({ children }) {
   const isAuth = localStorage.getItem('auth') === 'true'
-  return isAuth ? children : <Navigate to="/home" />
+  return isAuth ? children : <Navigate to="/" />
 }
